备份_纯python版本/Test/requirements.txt
altgraph==0.17
cffi==1.14.0
future==0.18.2
gevent==1.5.0
greenlet==0.4.15
labelImg==1.8.3
lxml==4.5.0
modbus-tk==1.1.0
numpy==1.18.1
opencv-python==4.2.0.32
pefile==2019.4.18
pycparser==2.20
PyInstaller==4.0.dev0+f877f3eb8a
PyQt5==5.13.0
PyQt5-sip==12.7.1
pyqt5-tools==5.13.0.1.5
pyserial==3.4
pywin32-ctypes==0.2.0
